NTBAD: split into NewTaskDialog
